Initial upload of skeleton on project
At this point Colorator is capable of generating 4 random colors at a time
with its hexadecimal identity. Four colors are spread over four DIVs of HTML.

Colors are defined by an array of hexadecimal characters [1, 2, 3, 4, 5, 6, 7,
8, 9, 0, a, b, c, d, e, f]. These characters are randomized by the rand()
function of PHP, and thus produces different identities every time the script
is requested.

There is a JavaScript function which disables the right click menu and there
is another function which reloads the page on right click.

It is an initial upload and many things are to be included. Colorator is
currently only tested in Chromium.